+++ packages/use-clipboard/__tests__/index.test.tsx
@@ -2,12 +2,12 @@ import React from 'react';
 import { cleanup, render, screen, fireEvent, act } from '@testing-library/react';
 import useClipboard from '../index';
 
-function Template({ interval }: { interval: number }) {
-  const [run, state] = useClipboard('copy text', interval);
+function Template({ interval, text }: { interval: number; text: string }) {
+  const [run, state] = useClipboard(interval);
 
   return (
     <>
-      <button onClick={run} data-testid="btn-run">
+      <button onClick={() => run(text)} data-testid="btn-run">
         run
       </button>
       <div data-testid="text-copied">{state.isCopied ? 'Yes' : 'No'}</div>
@@ -20,7 +20,7 @@ describe('test useClipboard', () => {
   afterEach(cleanup);
 
   it('when interval <= 0', () => {
-    render(<Template interval={0} />);
+    render(<Template interval={0} text="copy" />);
 
     expect(screen.getByTestId('text-copied').textContent).toEqual('No');
 
@@ -32,7 +32,7 @@ describe('test useClipboard', () => {
   it('when interval > 0', () => {
     jest.useFakeTimers();
     const interval = 1000;
-    render(<Template interval={interval} />);
+    render(<Template interval={interval} text="copy" />);
 
     expect(screen.getByTestId('text-copied').textContent).toEqual('No');