+++ packages/utils/index.ts
@@ -3,5 +3,6 @@ import sortDataList from './sortDataList';
 import stringToLower from './stringToLower';
 import getElement from './getElement';
 import keyboardMapper from './keyboardMapper';
+import isBrowser from './isBrowser';
 
-export { RafTimer, sortDataList, stringToLower, getElement, keyboardMapper };
+export { RafTimer, sortDataList, stringToLower, getElement, keyboardMapper, isBrowser };